:construction: Make the code more readable
