+++ src/utils/enum.ts
@@ -1,0 +1,14 @@
+/**
+ * export support platform enum
+ */
+export enum platformEnum {
+    window = 'win32',
+    linux = 'linux',
+    macOS = 'darwin'
+}
+
+
+/**
+ * defined the type of status that allSettled function returned
+ */
+export enum allSettledStatusEnum { fulfilled = 'fulfilled',  rejected = 'rejected'}
